+++ front/components/assistant/conversation/ConversationViewer.tsx
